feat: hub /planning con tabs Metas + Presupuestos
Replica la sección de Planificación de la versión web. Tab toggle
interno con segmented control (sin material-top-tabs). FAB apunta
dinámicamente a /savings/new o /budgets/new según tab activo.

Ambos listados quedan montados a la vez con display none/flex para
preservar scroll position al hacer toggle.

- components/savings/SavingsList.tsx: nuevo, lista reutilizable
- components/budgets/BudgetsList.tsx: nuevo, lista reutilizable
- app/(dashboard)/savings/index.tsx: refactor para usar SavingsList
- app/(dashboard)/budgets/index.tsx: refactor para usar BudgetsList
- app/(dashboard)/planning/: nuevo hub
- _layout.tsx: registrar planning con href: null
- QuickActionsSheet: reemplazar tiles 'Presupuestos' y 'Metas' por
  un solo tile 'Planificación' que abre el hub

Las rutas legacy /savings y /budgets siguen funcionando para deep
links y navegación desde pantallas de detalle.

Refs: #110
