+++ app/(dashboard)/_layout.tsx
@@ -108,6 +108,7 @@ export default function DashboardLayout() {
         <Tabs.Screen name="categories" options={{ href: null }} />
         <Tabs.Screen name="budgets" options={{ href: null }} />
         <Tabs.Screen name="savings" options={{ href: null }} />
+        <Tabs.Screen name="planning" options={{ href: null }} />
         <Tabs.Screen name="loans" options={{ href: null }} />
         <Tabs.Screen name="reminders" options={{ href: null }} />
       </Tabs>
