+++ app/(dashboard)/savings/index.tsx
@@ -1,39 +1,17 @@
+// app/(dashboard)/savings/index.tsx
+//
+// Pantalla dedicada de Metas de Ahorro. Mantiene header con back + FAB.
+// El contenido del listado lo provee <SavingsList /> (reutilizable, también
+// usado por el hub /planning).
 import { useCallback, useState } from 'react'
-import {
-  View,
-  Text,
-  FlatList,
-  TouchableOpacity,
-  ActivityIndicator,
-  RefreshControl,
-} from 'react-native'
+import { View, Text, TouchableOpacity } from 'react-native'
 import { router, useFocusEffect } from 'expo-router'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { useAuth } from '@/context/AuthContext'
 import { getSavingsGoals } from '@/lib/actions/savings.actions'
-import { ProgressBar } from '@/components/ui/ProgressBar'
-import { EmptyState } from '@/components/ui/EmptyState'
-import { ListLoadingSkeleton } from '@/components/ui/ListLoadingSkeleton'
+import { SavingsList } from '@/components/savings/SavingsList'
 import { toast } from '@/components/ui/Toast'
-import { formatCurrency } from '@/lib/utils/currency'
 import type { SavingsGoal } from '@/types/database.types'
-
-/** Devuelve "Faltan X días", "Vence hoy", "Vencida hace Y días", o null si no hay deadline. */
-function formatDeadline(deadline: string | null): {
-  text: string
-  overdue: boolean
-} | null {
-  if (!deadline) return null
-  const today = new Date()
-  today.setHours(0, 0, 0, 0)
-  const target = new Date(`${deadline}T00:00:00`)
-  const diffMs = target.getTime() - today.getTime()
-  const days = Math.round(diffMs / (1000 * 60 * 60 * 24))
-
-  if (days === 0) return { text: 'Vence hoy', overdue: false }
-  if (days > 0) return { text: `Faltan ${days} días`, overdue: false }
-  return { text: `Vencida hace ${Math.abs(days)} días`, overdue: true }
-}
 
 export default function SavingsScreen() {
   const { user } = useAuth()
@@ -62,19 +40,6 @@ export default function SavingsScreen() {
     }, [loadGoals])
   )
 
-  if (loading) {
-    return (
-      <SafeAreaView edges={['top']} className="flex-1 bg-bg">
-        <View className="flex-row items-center justify-between px-4 py-3 border-b border-border">
-          <View style={{ width: 60 }} />
-          <Text className="text-white text-base font-bold">Metas de ahorro</Text>
-          <View style={{ width: 60 }} />
-        </View>
-        <ListLoadingSkeleton />
-      </SafeAreaView>
-    )
-  }
-
   return (
     <SafeAreaView edges={['top']} className="flex-1 bg-bg">
       {/* Header */}
@@ -86,31 +51,15 @@ export default function SavingsScreen() {
         <View style={{ width: 60 }} />
       </View>
 
-      <FlatList
-        data={goals}
-        keyExtractor={(g) => g.id}
-        contentContainerStyle={
-          goals.length === 0 ? { flex: 1 } : { paddingVertical: 12 }
-        }
-        ListEmptyComponent={
-          <EmptyState
-            icon="🐖"
-            title="Sin metas"
-            description="Toca + para crear tu primera meta de ahorro"
-          />
-        }
-        renderItem={({ item }) => <GoalCard goal={item} />}
-        refreshControl={
-          <RefreshControl
-            refreshing={refreshing}
-            onRefresh={async () => {
-              setRefreshing(true)
-              await loadGoals()
-              setRefreshing(false)
-            }}
-            tintColor="#4F46E5"
-          />
-        }
+      <SavingsList
+        goals={goals}
+        loading={loading}
+        refreshing={refreshing}
+        onRefresh={async () => {
+          setRefreshing(true)
+          await loadGoals()
+          setRefreshing(false)
+        }}
       />
 
       {/* FAB */}
@@ -133,76 +82,5 @@ export default function SavingsScreen() {
         </Text>
       </TouchableOpacity>
     </SafeAreaView>
-  )
-}
-
-function GoalCard({ goal }: { goal: SavingsGoal }) {
-  const target = Number(goal.target_amount)
-  const current = Number(goal.current_amount)
-  const ratio = target > 0 ? current / target : 0
-  const remaining = Math.max(0, target - current)
-  const deadline = formatDeadline(goal.deadline)
-  const completed = goal.is_completed || ratio >= 1
-
-  return (
-    <TouchableOpacity
-      onPress={() => router.push(`/savings/${goal.id}`)}
-      activeOpacity={0.7}
-      className="mx-4 mb-3 bg-surface border border-border rounded-2xl p-4"
-    >
-      <View className="flex-row items-center mb-3">
-        <View className="flex-1">
-          <View className="flex-row items-center gap-2">
-            <Text className="text-white text-base font-semibold">
-              {goal.name}
-            </Text>
-            {completed ? (
-              <View className="px-2 py-0.5 rounded-full bg-green-500/15 border border-green-500/40">
-                <Text className="text-green-400 text-xs font-semibold">
-                  Completada
-                </Text>
-              </View>
-            ) : null}
-          </View>
-          {deadline ? (
-            <Text
-              className={
-                deadline.overdue
-                  ? 'text-red-400 text-xs mt-0.5'
-                  : 'text-muted text-xs mt-0.5'
-              }
-            >
-              {deadline.text}
-            </Text>
-          ) : null}
-        </View>
-        <Text className="text-green-400 text-base font-bold">
-          {Math.round(ratio * 100)}%
-        </Text>
-      </View>
-
-      <ProgressBar value={ratio} color="#10b981" />
-
-      <View className="flex-row justify-between mt-3">
-        <Text className="text-muted text-xs">
-          Ahorrado{' '}
-          <Text className="text-white text-sm">
-            {formatCurrency(current, goal.currency)}
-          </Text>
-        </Text>
-        <Text className="text-muted text-xs">
-          de{' '}
-          <Text className="text-white text-sm">
-            {formatCurrency(target, goal.currency)}
-          </Text>
-        </Text>
-      </View>
-
-      {!completed ? (
-        <Text className="text-muted text-xs mt-2">
-          Restante {formatCurrency(remaining, goal.currency)}
-        </Text>
-      ) : null}
-    </TouchableOpacity>
   )
 }
